int64 index default again, make R compatible

+++ darr/raggedarray.py
@@ -290,7 +290,7 @@ class RaggedArray:
 # FIXME empty arrayiterable
 # TODO describe int32 vs int64 index
 def asraggedarray(path, arrayiterable, dtype=None, metadata=None,
-                  accessmode='r+', indextype='int32', overwrite=False):
+                  accessmode='r+', indextype='int64', overwrite=False):
     path = Path(path)
     supportedindextypes = ('int8','uint8', 'int16', 'uint16', 'int32',
                            'uint32', 'int64')
@@ -345,7 +345,7 @@ def asraggedarray(path, arrayiterable, dtype=None, metadata=None,
 
 
 def create_raggedarray(path, atom=(), dtype='float64', metadata=None,
-                       accessmode='r+', indextype='int32', overwrite=False):
+                       accessmode='r+', indextype='int64', overwrite=False):
     if not hasattr(atom, '__len__'):
         raise TypeError(f'shape "{atom}" is not a sequence of dimensions.\n'
                         f'If you want just a list of 1-dimensional arrays, '
